Link up scraping results to local api endpoint
