+++ app/actions/images.js
@@ -1,0 +1,8 @@
+// import { scrapeImages } from '../helpers/scraperUtils';
+//
+//
+// export const scrapeImagesAction = () => {
+//   // return (dispatch) => {
+//   scrapeImages().then(data => console.log(data));
+//   // };
+// };
