+++ app/components/App.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-// import { Route, Redirect } from 'react-router-dom';
 import Header from './Header';
 
 class App extends Component {
