+++ app/reducers/index.js
@@ -3,8 +3,6 @@ import { routerReducer } from 'react-router-redux';
 import { imagesErrored, imagesLoading, images } from './images-reducers';
 
 const rootReducer = combineReducers({
-  imagesErrored,
-  imagesLoading,
   images,
   router: routerReducer(),
 });
