+++ package.json
@@ -5,7 +5,8 @@
   "scripts": {
     "build": "webpack --config ./webpack.config.js",
     "dev-server": "webpack-dev-server --hot --inline --content-base --history-api-fallback",
-    "start": "npm build && nodemon src/app.js"
+    "start": "npm build && nodemon src/app.js",
+    "scrape": "node src/helpers/scraperUtils.js"
   },
   "keywords": [],
   "author": "Lindsay Parker",
